requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
cycler==0.10.0
Django==2.1.4
dlib==19.16.99
future==0.17.1
h5py==2.8.0
idna==2.8
Keras==2.0.2
kiwisolver==1.0.1
matplotlib==3.0.2
numexpr==2.6.8
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.23.4
protobuf==3.6.1
pycparser==2.19
pyparsing==2.3.0
PySocks==1.6.8
python-dateutil==2.7.5
python-telegram-bot==11.1.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
scipy==1.1.0
six==1.12.0
tables==3.4.4
tensorflow==1.0.0
Theano==1.0.3
tqdm==4.28.1
urllib3==1.24.1
